add `ui/utils.rs` module
